refactor: move isAppIcon function into icon component
